+++ src/search/binarySearch/shipWithinDays.ts
@@ -1,5 +1,7 @@
 // https://leetcode-cn.com/problems/capacity-to-ship-packages-within-d-days/
 // 1011
+// https://leetcode-cn.com/problems/split-array-largest-sum/
+// 410
 export function shipWithinDays(weights: number[], day: number): number {
   let right = weights.reduce((prev, current) => current + prev, 0);
   let left = Math.max.apply(null, weights);